:recycle: refactor: Exception 개편(간단하게 구현)
