--- application/core-exception/src/main/java/com/fundy/application/exception/CoreExceptionType.java
@@ -1,5 +1,0 @@
-package com.fundy.application.exception;
-
-public enum CoreExceptionType {
-    DUPLICATE
-}
