--- application/core-exception/src/main/java/com/fundy/application/exception/DuplicateInstanceException.java
@@ -1,7 +1,0 @@
-package com.fundy.application.exception;
-
-public class DuplicateInstanceException extends CoreApplicationException {
-    protected DuplicateInstanceException(String message) {
-        super(message);
-    }
-}
