+++ application/core-exception/src/main/java/com/fundy/application/exception/custom/CoreApplicationException.java
@@ -1,4 +1,4 @@
-package com.fundy.application.exception;
+package com.fundy.application.exception.custom;
 
 public class CoreApplicationException extends RuntimeException {
     protected CoreApplicationException(String message) {
